code: move errors to api to have them in one place
